refactor: 리뷰 반영으로 AnalysisController 관련 DTO, Repo, Service 내부 수정
- AnalysisController에 있던 Service들을 AnalysisService로 이동하여 Controller를 단순화함
- DTO인 ArticlesAnalysisResponse, CommentsAnalysisResponse의 필드 자료형을 연산 결과대로 수정
- AnalysisService에서 기타 Service들을 이용하여 응답에 필요한 값을 만들어내도록 하며, DB 접근을 최소화하는 로직으로 변경,
  Member가 작성한 Comment의 총 개수를 반환하는 메서드 이름도 직관적으로 고침
- ArticleEmotionService에서 AnalysisService가 이용하는 메서드 로직들을 수정
- CommentService에서 AnalysisService가 이용하는 메서드의 반환값을 int로 변환하지 않고 반환 (int로 변환해야 할 필요가 없음)
- ArticleEmotionRepository에서 주어진 Article들에 해당하는 ArticleEmotion을 찾아오도록 메서드 변경
  기존 쿼리 메서드에 비해 로직의 흐름을 이해하기 쉽도록 함

+++ src/main/java/com/saebyeok/saebyeok/dto/CommentsAnalysisResponse.java
@@ -8,5 +8,5 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @Getter
 public class CommentsAnalysisResponse {
-    private int totalCommentsCount;
+    private Long totalCommentsCount;
 }
